2365 payment history refactor (#9037)
We are adding an adapter to sit between the BGS payment service's response and the serializer. The mobile api then subclasses the adapter as we include fields (id and date) that the core API does not. The mobile API will also only display payments while the core API displays payments and return payments. I added a use case to the controller specs (plus a vcr cassette) for when there are both payments and return payments.