Check for COE application submissions without an edipi (#10811)
We have been getting 404s when hitting LGY's put_application endpoint on staging. See https://github.com/department-of-veterans-affairs/va.gov-team/issues/47435. It is currently unclear why. Since a 404 generally indicates that the server cannot find the requested resource, it is possible that we are not including an edipi with our requests, and this omission is causing LGY to return a 404. This commit logs an error to sentry if that is the case. Once deployed to staging, if we receive a missing edipi sentry error at the same time as a 404 error, then we'll know that a missing edipi is likely responsible for our 404s. If not, then not.