Fix LGY put_application endpoint error logging. (#10897)
In https://github.com/department-of-veterans-affairs/vets-api/pull/10847/, I tried expanding the scope of errors that we logged from LGY. Specifically, I tried rescuing `Faraday::Error`s under the assumption that `Faraday::ClientError`s and `Faraday::ServerError`s both inherited from `Faraday::Error`s. Unfortunately, that is not true. It turns out that 400 and 500 errors manifest as `Faraday::ClientError`s and `Faraday::Error`s are not instances of `Faraday::Error`. This PR rescues the right errors, so that they can be logged in Sentry.