Hardcoded COE params hotfix. (#11262)
In https://github.com/department-of-veterans-affairs/vets-api/pull/11237, I failed to consider that if the vet did not indicate any prior loans, `relevantPriorLoans` would be `nil` rather than an empty list. Consequently, `.any?` would throw an error when a vet did not indicate any prior loans, leading to a 500 error and failed form submission. This PR accounts for this case.