[50438] Display only notification letters on the COE status page. (#11335)
LGY's documents endpoint returns all documents for a vet without discriminating between vet-uploaded supporting documents and provider-issued notification letters. But we need to display only notification letters on the COE status page, and not vet-uploaded supporting documents. That requires us to distinguish between the two. In order to make that distinction, I am adding an "[ATTACHMENT]" tag to the filename and description of every vet-uploaded file. Currently, there is no reasonable alternative solution to this one.