Simplify notification letter filtering for COE status page. (#11338)
* Simplify notification letter filtering for COE status page.

This is a follow up from https://github.com/department-of-veterans-affairs/vets-api/pull/11335. There, I added the "[ATTACHMENT]" tag to both the fileName and description because I wasn't sure whether POSTing these parameters would result in those fields being returned in our subsequent GET /documents response. Testing on staging confirmed that the fileName and description in our POST request does translate to the mimeType(†) and description in the GET /documents response. This means that I only need to add a tag to one of these fields. It doesn't really matter which one I choose, but I'm choosing to add the tag to `description` because it is used less than `fileName`.