Adjust when json middleware runs (#12290)
* Adjust response json parsing content-type

We use the Faraday json middleware to automatically parse json responses
to Ruby hashes. This wasn't working when trying to hit an endpoint that
downloads a PDF. Even though Faraday docs for this say it defaults to
only parsing `application/json` content-types, there was a subtle way in
which the lighthouse api returns its content type: `Content-Type:
application/json;charset-UTF8`. The default for the json middleware was
set to `/\bjson$/` which only matches if the content type _ends_ in
`json`.