Set ssn to User#ssn if malformed in SubmitForm674Job. (#12410)
In https://github.com/department-of-veterans-affairs/vets-api/pull/12379, I set up logging to catch if we were sending malformed SSNs in our `vnp_person_create` call to BGS. That work was predicated upon the assumption that when BGS returned a failed SQL statement with an SSN of "********", BGS was redacted the actual SSN we sent over. After deploying this work, I learned that we were literally sending SSNs equal to the string "********" to BGS.

The SSN we send originates from User#ssn, in BGS::DependentService. It is passed to SubmitForm686cJob, which makes a successful request to BGS (i.e. the SSN at this stage is a valid 9-digit SSN). That SSN is then passed to SubmitForm674Job; however, by the time we are sending it to BGS, it has transformed into the string "********". I have no idea why this is happening, even after spending several hours investigating. I'm going to continue that investigation, but in the meantime, in this PR, I'm going to try to "refresh" the ssn by calling User#ssn once again, if the SSN is malformed before we send it to BGS.