Handle special characters in names before sending data to BGS. (#13135)
In https://github.com/department-of-veterans-affairs/vets-api/pull/12968, I wrote code to normalize name and address data in the form data we received from VA.gov, before sending it to BGS. However, I failed to do the same for name data from the `User` object included in the payload to BGS's `insert_benefit_claim` endpoint. In this PR, I've abstracted the normalization code into a helper module, and used it for the `User` name data, as well as the form data received from VA.gov.